+++ src/ordering-co/ordering-co.service.ts
@@ -2,18 +2,18 @@ import { ForbiddenException, HttpException, Injectable } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import axios from 'axios';
 import { LoginDto } from 'src/auth/dto/auth';
-import { OrderingSignInResponseDto } from './dto/ordering-co.dto';
+import { OrderingSignInResponseDto } from './dto/signin-response.dto';
 
 @Injectable()
 export class OrderingCoService {
   constructor(private configService: ConfigService) {}
   public validateWebhookSecret(checkSecret: string): boolean {
-    let webhookSecret = this.configService.get<string>('ordering_co_webhook_secret') as string;
+    let webhookSecret = this.configService.get<string>('orderingco.webhook.secret') as string;
     return webhookSecret === checkSecret;
   }
 
   public getWebhookSecret(): string {
-    return this.configService.get<string>('ordering_co_webhook_secret') as string;
+    return this.configService.get<string>('orderingco.webhook.secret') as string;
   }
 
   async signIn(loginDto: LoginDto): Promise<OrderingSignInResponseDto> {
